Refactor class packages
