+++ src/main/java/com/snowypeaksystems/mobportals/listeners/CommandListener.java
@@ -1,7 +1,8 @@
-package com.snowypeaksystems.mobportals;
+package com.snowypeaksystems.mobportals.listeners;
 
 import static com.snowypeaksystems.mobportals.messages.Messages.gm;
 
+import com.snowypeaksystems.mobportals.MobPortals;
 import com.snowypeaksystems.mobportals.exceptions.PermissionException;
 import java.io.IOException;
 import java.util.ArrayList;
